add comment function part1

+++ src/main/java/life/east/community/controller/PublishController.java
@@ -2,7 +2,6 @@ package life.east.community.controller;
 
 import com.mysql.jdbc.StringUtils;
 import life.east.community.dto.QuestionDTO;
-import life.east.community.mapper.QuestionMapper;
 import life.east.community.model.Question;
 import life.east.community.model.User;
 import life.east.community.service.QuestionService;
@@ -28,7 +27,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.findById(id);
         model.addAttribute("title", question.getTitle());
@@ -58,7 +57,7 @@ public class PublishController {
             @RequestParam(name = "title",required = false) String title,
             @RequestParam(name = "description",required = false) String description,
             @RequestParam(name = "tag",required = false) String tag,
-            @RequestParam(name = "id",required = false) Integer id,
+            @RequestParam(name = "id",required = false) Long id,
             HttpServletRequest request,
             Model model) {
